requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
Pillow==9.4.0
pytesseract==0.3.10